% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findroot.R
\name{findroot}
\alias{findroot}
\title{One-to-one mapping relating expectiles and quantiles}
\usage{
findroot(X, Y, omega, alpha, p = 1)
}
\arguments{
\item{X}{The covariate data values.}

\item{Y}{The response data values.}

\item{omega}{Numeric vector of level between 0 and 1 where 0.5 corresponds
to the mean.}

\item{alpha}{Alpha value corresponding to the omega value given by the
one-to-one mapping}

\item{p}{The order of the local polynomial estimator. In default setting,
\code{p=1}.}
}
\value{
\code{\link{findroot}} returns the one-to-one mapping
}
\description{
Ont-to-one mapping (see Proposition 1 in Adam and Gijbels (2021a))
}
\references{
{

Adam, C. and Gijbels, I. (2021a). Local polynomial expectile regression.
Annals of the Institute of Statistical Mathematics doi:10.1007/s10463-021-00799-y.

}
}
